feat(functions): consolidate function calling settings and remove deprecated files
